% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crl.mat}
\alias{crl.mat}
\title{Continuation ratio logit transformation}
\format{
A data frame with 58 observations and 11 variables; rows represent
years, columns represent ages and values are crl at that age and year
}
\usage{
crl.mat
}
\description{
The age composition data are fit using the continuation ratio
logit (crl) transformation (see Perreault et al., 2020) for details.
}
\keyword{datasets}
