% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cw.mat}
\alias{cw.mat}
\title{Maturity at age}
\format{
A data frame with 58 observations and 12 variables; rows represent
years, columns represent ages and values are weight in kg.
}
\usage{
cw.mat
}
\description{
Estimated commercial catch weights for 3LNO American plaice
}
\keyword{datasets}
