% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FRV.vec}
\alias{FRV.vec}
\title{Fall survey indices}
\format{
A data frame with 390 and 4 variables
\describe{
\item{Year}{survey year}
\item{Age}{Age of }
\item{index}{Standardized survey index}
\item{survey}{Indicator of survey type}
}
}
\usage{
FRV.vec
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
