% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_LI.R
\name{gfunc}
\alias{gfunc}
\title{To calculate influence function}
\usage{
gfunc(w, theta, mfits)
}
\arguments{
\item{w}{weight for perturbation (set to 0 for null perturbation)}

\item{theta}{vector of fixed effects}

\item{mfits}{object from make_fit}
}
\description{
To calculate influence function
}
