% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{index}
\alias{index}
\title{3LNO American plaice survey indices}
\format{
A data frame with 840 observations and 12 variables
\describe{
\item{Year}{survey year}
\item{Age}{survey age in years}
\item{index}{survey abundance index in millions}
\item{survey}{Type of survey; Fall/Spring}
\item{fs}{Fraction of year that survey takes places}
\item{iyear}{year indicator}
\item{iage}{age indicator}
\item{isurvey}{survey indicator}
\item{surv_age}{identifier survey and age combination}
\item{isd}{survey/age identifier}
\item{surv_year}{identifier survey and year combination}
\item{is_year}{survey/year identifier}
}
}
\usage{
index
}
\description{
A dataset containing the survey data for 3LNO American plaice
}
\keyword{datasets}
