% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land}
\alias{land}
\title{Commercial landings}
\format{
A data frame with 58 observations and 2 variables
\describe{
\item{Landings}{landings estimates (t)}
\item{Year}{landed year}

}
}
\usage{
land
}
\description{
Nominal landings (t) of American plaice for NAFO divisions 3LNO
}
\keyword{datasets}
