% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_curv_plots.R
\name{make.curv.plot}
\alias{make.curv.plot}
\title{make.curv.plots: plot curvature}
\usage{
make.curv.plot(LocInf, curv)
}
\arguments{
\item{LocInf}{object from \code{\link{make.LI}}}

\item{curv}{object from \code{\link{make.curv}}}
}
\value{
a list containing  ggplot
\describe{
\item{\code{$Curv_vs_LI_plot}}{plots curvature vs influence slopes}
}
}
\description{
Plot output from curvature.
}
\examples{
\dontrun{
curv.plots <- make.curv.plots(mfits, curv)
}
}
