% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_curv.R
\name{make.curv}
\alias{make.curv}
\title{make.curv: calculate curvature}
\usage{
make.curv(mfits, LocInf, pert = NULL, tol = NULL)
}
\arguments{
\item{mfits}{object returned from \code{\link{make.fit}}}

\item{LocInf}{object returned from \code{\link{make.LI}}}

\item{pert}{(optional) vector, to manually set which perturbations to run. Runs all by default.}

\item{tol}{to set curvature tolerance. Default \code{h=0.10}.}
}
\description{
Calculates curvature for APAM for individual, age group and year group slopes for full model fit.
Note that this functions currently does not calculate curvature for individual
data components.
}
\examples{
\dontrun{
curv<- make.curv(mfits,LocInf)
}
}
