% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fit.R
\name{make.fit}
\alias{make.fit}
\title{Prepare parameters for APAM}
\usage{
make.fit(
  data,
  map,
  parameters,
  do.sd = TRUE,
  do.Zresid = TRUE,
  do.NS = TRUE,
  do.hess = FALSE,
  NS = 4
)
}
\arguments{
\item{data}{input from make_tmb_data}

\item{map}{input from make_map}

\item{parameters}{input from make_parameters}

\item{do.sd}{get sdreport}

\item{do.Zresid}{get Z residuals}

\item{do.NS}{run extra Newton steps (default 4)}

\item{do.hess}{calculate hessian}

\item{NS}{default 4 extra Newton step}
}
\description{
Prepares parameters for use in APAM
}
