% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_LI.R
\name{make.LI}
\alias{make.LI}
\title{make.LI: run local influence diagnostics for APAM.}
\usage{
make.LI(mfits, pert = NULL, all = TRUE, age = FALSE, year = FALSE)
}
\arguments{
\item{mfits}{object returned from \code{\link{make.fit}}}

\item{pert}{(optional) vector, to manually set which perturbations to run. Runs all by default. See details.}

\item{all}{T/F, to run individual (i.e. age and year) perturbations. Default = \code{TRUE}.}

\item{age}{T/F, to run age group perturbations. Default = \code{FALSE}.}

\item{year}{T/F, to run year group perturbations. Default = \code{FALSE}.}
}
\description{
To run local influence diagnostics for full model and individual data components.
}
\examples{
\dontrun{
LI <- make.LI(fit)

#to run age group perturbations
LI_age <- make.LI(fit,age=T)
}
}
