% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.R
\name{make.map}
\alias{make.map}
\title{Prepare map for APAM}
\usage{
make.map(
  data,
  setmap = NULL,
  crl.block = T,
  no.pe = F,
  no.logits = F,
  no.Flogits = F
)
}
\arguments{
\item{data}{make_tmb_data object}

\item{setmap}{can be used to change default map}

\item{crl.block}{T/F, turn on/off crl sd parm split pre/post 1990}

\item{no.pe}{T/F. turn on/off process errors}

\item{no.logits}{T/F, turn on/off logit estimation}

\item{no.Flogits}{T/F, turn on/off Flogit estimation}
}
\description{
Prepares map for use in APAM. Takes make_tmb_data object as input
}
