% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots.R
\name{make.plots}
\alias{make.plots}
\title{make.plots: plot output from APAM}
\usage{
make.plots(mfits)
}
\arguments{
\item{mfits}{object from \code{\link{make.fit}}}
}
\value{
a list containing all ggplots
\describe{
\item{\code{$indexfit1}}{observed vs predicted survey indices at age for ages 1-7}
\item{\code{$indexfit2}}{observed vs predicted survey indices at age for ages 8-15}
\item{\code{$resid_index}}{survey Z residuals}
\item{\code{$pop_process}}{estimated population processes (recruitment, average fishing mortality rates,
spawning stock biomass and population abundance)}
\item{\code{$index_bplot}}{bubble plot of survey index residuals}
\item{\code{$crl_fit}}{observed vs predicted crl fits}
\item{\code{$resid_crl}}{crl Z residuals}
\item{\code{$crl_bplot}}{bubble plot of crl residuals}
\item{\code{$landings}}{estimated model landings with upper/lower bounds}
\item{\code{$landratio}}{estimated landings ratios}
\item{\code{$F_at_age}}{model predicted fishing mortality rates at age}
\item{\code{$log_F_dev}}{model predicted log fishing mortality rate deviations}
\item{\code{$process_errors}}{model predicted process errors}
\item{\code{$pe_bplot}}{bubble plot of model predicted process errors}
\item{\code{$catchability}}{estimated survey cathcability}
}
}
\description{
Plots output from APAM.
}
\examples{
\dontrun{
plots<- make.plots(mfits)
}
}
