% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_profile_plots.R
\name{make.profile.plots}
\alias{make.profile.plots}
\title{Plot output from \code{\link{make_profile}} object}
\usage{
make.profile.plots(profile)
}
\arguments{
\item{prof}{object from \code{\link{make_profile}}}
}
\description{
Outputs profile plots
}
