% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matur.mat}
\alias{matur.mat}
\title{Maturity at age}
\format{
A data frame with 59 observations and 12 variables; rows represent
years, columns represent ages and values are proportion mature
}
\usage{
matur.mat
}
\description{
Estimated proportion mature for 3LNO American plaice
}
\keyword{datasets}
