% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sw.mat}
\alias{sw.mat}
\title{Stock weights}
\format{
A data frame with 58 observations and 15 variables; rows represent
years, columns represent ages and values are weight in kg.
}
\usage{
sw.mat
}
\description{
Stock weights estimated from from a spatiotemporal biphasic vonBertalanffy
growth model. accounted for the length-stratified age sampling design.
The 3LNO stock weights were combined for each division
by weighting the values for each division by the average
abundance index at age during 1975–2017. Stock weights
prior to 1975 were fixed at the mean values for 1975–77.
}
\keyword{datasets}
