% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_LI.R
\name{dFgrad_dtheta}
\alias{dFgrad_dtheta}
\title{Helper functions for local influence diagnostics}
\usage{
dFgrad_dtheta(w, i, mfits)
}
\arguments{
\item{w}{weight for perturbation (set to 0 for null perturbation)}

\item{i}{inidcator for observation weight}

\item{mfits}{object from make_fit}
}
\description{
#function to compute Del in Eq. 22 of Perreault & Cadigan
#'@useDynLib APAM
}
