% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_LI2.R
\name{gfunc}
\alias{gfunc}
\title{gfunc: helper function to return model fit}
\usage{
gfunc(w, theta, mfits)
}
\arguments{
\item{w}{set weight for perturbation (0 for null perturbation).}

\item{theta}{vector of fixed effect parameters}

\item{mfits}{object returned from \code{\link{make.fit}}}
}
\description{
Influence function for local influence diagnostics.
}
