% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_LI_plots.R
\name{make.LI.plots}
\alias{make.LI.plots}
\title{make.LI.plots: plot output from local influence diagnostics}
\usage{
make.LI.plots(LocInf)
}
\arguments{
\item{LocInf}{object from \code{\link{make.LI}}}
}
\value{
a list containing  ggplot
\describe{
\item{\code{$LI_plot}}{local influence plot}
}
}
\description{
Plot outputs from APAM local influence diagnostics.
}
\examples{
\dontrun{
LI.plots <- make.LI.plots(LocInf)
}
}
