% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_LI.R
\name{make.LI}
\alias{make.LI}
\title{To run local influence diagnostics for individual data components}
\usage{
make.LI(mfits, pert = NULL, full = NULL, all = TRUE, age = FALSE, year = FALSE)
}
\arguments{
\item{mfits}{object from make_fit}

\item{pert}{define which perturbations to run; default=all (for all ages/years)}

\item{full}{if running for data source, need to specify full model fit for Del calculation}

\item{all}{to run individual perturbations}

\item{age}{to run age group perturbations}

\item{year}{tp run year group perturbations}
}
\description{
To run local influence diagnostics for individual data components
}
