% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_parm.R
\name{make.parm}
\alias{make.parm}
\title{Prepare parameters for APAM}
\usage{
make.parm(
  data,
  parmap,
  parameters = NULL,
  parmL = NULL,
  parmU = NULL,
  no.pe = F,
  no.logits = F,
  no.Flogits = F
)
}
\arguments{
\item{data}{input from make_tmb_data}

\item{parmap}{input from make_map}

\item{parameters}{can redefine starting parameter values}

\item{parmL}{can redefinie lower parameter bound}

\item{parmU}{can redefinie upper parameter bound}

\item{no.pe}{turn on/off pe, i.e fix pe sd to 0.01}

\item{no.logits}{turn on/off logits}

\item{no.Flogits}{turn on/off Flogit estimation}
}
\description{
Prepares parameters for use in APAM
}
