% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_parm.R
\name{make.parm}
\alias{make.parm}
\title{Prepare parms for APAM}
\usage{
make.parm(
  data,
  map,
  parms = NULL,
  parmsL = NULL,
  parmsU = NULL,
  no.pe = F,
  no.logits = F,
  no.Flogits = F
)
}
\arguments{
\item{data}{input from make_tmb_data}

\item{map}{input from make_map}

\item{parms}{can redefine starting parameter values}

\item{parmsL}{can redefine lower parameter bound}

\item{parmsU}{can redefine upper parameter bound}

\item{no.pe}{turn on/off pe, i.e fix pe sd to 0.01}

\item{no.logits}{turn on/off logits}

\item{no.Flogits}{turn on/off Flogit estimation}
}
\description{
Prepares parms for use in APAM
}
