% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_profile_plots.R
\name{make.profile.plots}
\alias{make.profile.plots}
\title{make.profile.plots: plot output from profile likelihoods}
\usage{
make.profile.plots(prof)
}
\arguments{
\item{prof}{object from \code{\link{make.profile}}}
}
\value{
a list containing all ggplots
\describe{
\item{\code{$nll_plot}}{profile likelihoods vs M perturbations}
\item{\code{$nll_dev_plot}}{profile likelihoods deviations vs M perturbations}
}
}
\description{
Plot outputs from APAM profile likelihoods
}
\examples{
\dontrun{
prof.plots<- make.profile.plots(prof)
}
}
