% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_profile.R
\name{make.profile}
\alias{make.profile}
\title{calculate profile likelihoods from Perreault & Cadigan 2021}
\usage{
make.profile(mfits, delta_M = NULL)
}
\arguments{
\item{mfits}{input from make_fit}

\item{delta_M}{to manually define delta M; default delta_M <- seq(-0.15,0.40, by = 0.01)}
}
\description{
Calculate profile likelihoods
}
