% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tmb_data.R
\name{make.tmb.data}
\alias{make.tmb.data}
\title{Prepare data inputs for APAM}
\usage{
make.tmb.data(
  do.retro = FALSE,
  retro.year = NULL,
  M.split = TRUE,
  M.matrix = NULL,
  C.bounds = NULL,
  sdL = NULL,
  d = NULL,
  SS = T
)
}
\arguments{
\item{do.retro}{T/F, turns on/off option to do retros}

\item{retro.year}{end year for retro.}

\item{M.split}{T/F, turns on/off M increase for years 1989-1996}

\item{M.matrix}{can redefine M input matrix;}

\item{sdL}{default landings sd fixed at 0.05;}

\item{d}{matrix of 0s; can be changed for influence diagnostics}

\item{C.bound}{defines landings upper bounds; default uncertainty(high=2,low = 1.2, moderate=1.5)}
}
\description{
Prepares data for use in APAM
}
